Add useUserForm hook, fixed styles
