+++ src/components/OrderRow/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
+import { Box, Typography } from '@mui/material';
 
 import { COLORS } from '@/constants/colors';
 import { FONT } from '@/constants/font';
-import { Box, Typography } from '@mui/material';
 
 import CustomCheckbox from '../Checkbox';
-import DriverAvatar from '../DriverAvatar';
+import DriverAvatar from '../UserAvatar';
 
 interface OrderRowProps {
   date: string;
