+++ src/pages/Orders/components/OrderItem/OrderItem.tsx
@@ -1,23 +1,23 @@
 import { FC } from 'react';
-import { Box, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
+import { Box, Typography } from '@mui/material';
 
-import { OrderItemInterface } from '@/pages/Orders/components/OrderItem/types';
-import { MONTHS } from '@/constants/moths';
-import DriverAvatar from '@/components/DriverAvatar';
 import Status from '@/components/Status';
-import {
-  orderRow,
-  customerBlock,
-  collectionDateStyles,
-  collectionAddressStyles,
-  routeBlock,
-  luggageStyles,
-  statusBlock,
-  collectionTimeStyles,
-  popup,
-} from '@/pages/Orders/components/OrderItem/styles';
+import DriverAvatar from '@/components/UserAvatar';
+import { MONTHS } from '@/constants/moths';
 import { OrderStatuses } from '@/interfaces/interfaces';
+import {
+  collectionAddressStyles,
+  collectionDateStyles,
+  collectionTimeStyles,
+  customerBlock,
+  luggageStyles,
+  orderRow,
+  popup,
+  routeBlock,
+  statusBlock,
+} from '@/pages/Orders/components/OrderItem/styles';
+import { OrderItemInterface } from '@/pages/Orders/components/OrderItem/types';
 
 const OrderItem: FC<OrderItemInterface> = ({
   collectionDate,
