+++ src/pages/TeamManagment/UserForm/useUserForm.tsx
@@ -1,0 +1,147 @@
+import { useState } from 'react';
+import { useForm } from 'react-hook-form';
+import { useTranslation } from 'react-i18next';
+import { useSelector } from 'react-redux';
+import { toast } from 'react-toastify';
+
+import { RootState } from '@/store/store';
+import axiosInstance from '@/utils/axiosInstance';
+import { getFirstName, getSecondName } from '@/utils/nameUtils';
+
+import { DecodedToken, UserFormInputs, UserFormProps } from '../types';
+
+export const useUserForm = ({
+  mode,
+  fetchUsers,
+  userId,
+  userData,
+  addUserToList,
+}: UserFormProps) => {
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [role, setRole] = useState(userData?.role || '');
+  const { t } = useTranslation();
+  const authToken = useSelector((store: RootState) => store.auth.token);
+
+  const {
+    register,
+    handleSubmit,
+    setValue,
+    reset,
+    formState: { errors },
+  } = useForm<UserFormInputs>({
+    defaultValues:
+      mode === 'update'
+        ? {
+            fullName: userData?.fullName || '',
+            email: userData?.email || '',
+            phoneNumber: userData?.phoneNumber || '',
+          }
+        : {
+            fullName: '',
+            email: '',
+            phoneNumber: '',
+            role: '',
+          },
+  });
+
+  const closeModal = (): void => {
+    setIsModalOpen(false);
+    reset();
+  };
+
+  const decodeToken = (token: string): DecodedToken | null => {
+    try {
+      const decodedToken = JSON.parse(atob(token.split('.')[1]));
+      return decodedToken;
+    } catch (error) {
+      console.error('Error decoding token:', error);
+      return null;
+    }
+  };
+
+  const sendData = async (formData: UserFormInputs) => {
+    try {
+      const decodedToken = authToken ? decodeToken(authToken) : null;
+
+      if (!decodedToken) {
+        console.error('Invalid token');
+        toast(t('settings.message.invalidToken'), { type: 'error' });
+        return;
+      }
+
+      const companyId = decodedToken.company_id.id;
+
+      if (!companyId) {
+        console.error('Company ID is missing from the token');
+        toast(t('settings.message.noCompanyId'), { type: 'error' });
+        return;
+      }
+
+      const transformedData = {
+        full_name: formData.fullName,
+        email: formData.email,
+        phone_number: formData.phoneNumber,
+        ...(mode === 'create' && { role: role.toLowerCase() }),
+        ...(mode === 'create' && { company_id: companyId }),
+      };
+
+      const url =
+        mode === 'create'
+          ? role === 'dispatcher'
+            ? '/dispatcher'
+            : '/driver'
+          : userData?.role === 'dispatcher'
+            ? `/dispatcher/${userId}`
+            : `/driver/${userId}`;
+
+      const method = mode === 'update' ? 'patch' : 'post';
+      const response = await axiosInstance[method](url, transformedData, {
+        headers: {
+          'Content-Type': 'application/json',
+          authorization: authToken || '',
+        },
+      });
+
+      const successMessage =
+        mode === 'update'
+          ? t('settings.message.updateSuccess')
+          : t('settings.message.addSuccess');
+      toast(successMessage, { type: 'success' });
+
+      if (mode === 'create') {
+        const newUser = response.data;
+
+        const adaptedUser = {
+          id: newUser.id,
+          role: newUser.role,
+          email: newUser.email,
+          phoneNumber: newUser.phone_number,
+          firstName: getFirstName(newUser.full_name),
+          lastName: getSecondName(newUser.full_name),
+        };
+
+        addUserToList(adaptedUser);
+      } else {
+        fetchUsers();
+      }
+
+      closeModal();
+    } catch (error) {
+      console.error(error);
+      toast(t('settings.message.error'), { type: 'error' });
+    }
+  };
+
+  return {
+    isModalOpen,
+    setIsModalOpen,
+    role,
+    setValue,
+    setRole,
+    register,
+    handleSubmit,
+    errors,
+    closeModal,
+    sendData,
+  };
+};
