+++ src/utils/nameUtils.ts
@@ -1,0 +1,7 @@
+export const getFirstName = (fullName: string): string => {
+  return fullName.split(' ')[0] || '';
+};
+
+export const getSecondName = (fullName: string): string => {
+  return fullName.split(' ').slice(1).join(' ') || '';
+};
